+++ pyrate_limiter/abstracts/__init__.py
@@ -1,3 +1,7 @@
+from .algorithm import Algorithm as Algorithm
+from .algorithm import Decision as Decision
+from .algorithm import LogAlgorithm as LogAlgorithm
+from .algorithm import SlidingWindowLog as SlidingWindowLog
 from .bucket import AbstractBucket as AbstractBucket
 from .bucket import BucketFactory as BucketFactory
 from .rate import Duration as Duration
@@ -6,6 +10,10 @@ from .rate import RateItem as RateItem
 from .wrappers import BucketAsyncWrapper as BucketAsyncWrapper
 
 __all__ = [
+    "Algorithm",
+    "Decision",
+    "LogAlgorithm",
+    "SlidingWindowLog",
     "AbstractBucket",
     "BucketFactory",
     "Duration",